Move referrals to utils
