+++ src/config/referrals.ts
@@ -1,0 +1,9 @@
+export const MAX_REFERRAL_CODE_LENGTH = 20;
+export const REFERRAL_CODE_REGEX = /^\w+$/; // only number, string and underscore is allowed
+
+export const TIER_DISCOUNT_INFO = {
+  0: 5,
+  1: 10,
+  2: 10,
+};
+
