+++ src/views/Referrals/AccountBanner.js
@@ -5,12 +5,11 @@ import {
   formatAmount,
   shortenAddress,
   USD_DECIMALS,
-  copyReferralCode,
-  getTierIdDisplay,
-  TIER_DISCOUNT_INFO,
 } from "../../Helpers";
 import CopyIcon from "../../img/copy.svg";
 import Tooltip from "../../components/Tooltip/Tooltip";
+import { copyReferralCode, getTierIdDisplay } from "../../utils/referrals";
+import { TIER_DISCOUNT_INFO } from "../../config/referrals";
 
 export default function AccountBanner(props) {
   const {
